requirements.txt
pdfplumber
markdown2
python-docx
Pillow
PyMuPDF
pytesseract
tqdm